fix: reduce reactive chain
