+++ apps/delivery-options/src/composables/index.ts
@@ -1,7 +1,6 @@
 export * from './events';
 export * from './useActiveCarriers';
 export * from './useBreakpoints';
-export * from './useCapabilitiesRequestParams';
 export * from './useDateFormat';
 export * from './useFeatures';
 export * from './useFormatDistance';
@@ -11,7 +10,6 @@ export * from './useOptionsGroupedByCarrier';
 export * from './usePickupLocation';
 export * from './usePickupLocationsMap';
 export * from './useProvideElementWidth';
-export * from './useSharedCapabilities';
 export * from './useResolvedCarrier';
 export * from './useResolvedDeliveryDates';
 export * from './useResolvedDeliveryMoments';
@@ -20,5 +18,6 @@ export * from './useResolvedPickupLocations';
 export * from './useSelectedDeliveryMoment';
 export * from './useSelectedPickupLocation';
 export * from './useSelectedValues';
+export * from './useSharedCapabilities';
 export * from './useShipmentOptionsOptions';
 export * from './useTimeRange';
